LLaMA/requirements_pip.txt
absl-py==2.1.0
accelerate==1.0.1
adamp==0.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.8.0
contourpy==1.3.0
cycler==0.12.1
DataProperty==1.1.0
datasets==2.21.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock @ file:///croot/filelock_1744281381737/work
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2023.6.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.11
importlib_metadata==8.6.1
importlib_resources==6.5.2
intel-cmplr-lib-ur==2025.2.1
intel-openmp==2025.2.1
interegular==0.3.3
ipykernel==6.28.0
ipython==8.18.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.10
lm_eval==0.4.8
lmflow==0.0.1.dev0
loguru==0.7.3
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral_common==1.5.3
mkl==2025.2.0
mkl-service==2.4.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.15
muon-optimizer @ git+https://github.com/KellerJordan/Muon@f90a42b28e00b8d9d2d05865fe90d9f39abcbcbd
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.56.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
outlines==0.0.46
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pathvalidate==3.2.3
pdf2image==1.17.0
peft==0.15.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
portalocker==3.1.1
powerlaw==1.5
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==0.0.1
pyarrow==19.0.1
pyarrow-hotfix==0.7
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyHessian @ git+https://github.com/amirgholami/PyHessian.git@de8cf239ace21c7754ee599255897ebf3eed19cb
pyparsing==3.2.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-ranger==0.1.1
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
ray==2.42.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
ruff==0.9.6
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.45.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tbb==2022.2.0
tcmlib==1.4.0
tcolorpy==0.1.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorly==0.9.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==0.4.5
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.8.0+cu128
torch-optimizer==0.3.0
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.46.1
triton==3.4.0
trl==0.12.0
typepy==1.3.4
typer==0.15.1
typing_extensions==4.15.0
tzdata==2025.1
umf==0.11.0
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.6.4.post1
wandb==0.17.3
watchfiles==1.0.4
wcwidth==0.2.13
websockets==12.0
weightwatcher==0.7.5.5
Werkzeug==3.1.3
word2number==1.1
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
